Refactored location of css files in one folder within lib - This is so as the site grows well into the future, the app folder wont bloat and distract from the pages tsx General - Moved css files to scss folder within lib - Added new path for scss folder - Changed import paths in files that used css jobCard css - Added new css file for jobCard component that takes the relevant css from home css global css - Changed shadow color to black
