+++ website/src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./root.scss";
+import "@/scss/layout/root.scss";
 
 const inter = Inter({ subsets: ["latin"] });
