+++ website/src/app/projects/starwarsapi/page.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
 import LogoRow from "@/lib/components/logoRow";
 import { starLogoData } from "@/lib/data/logoData";
-import styles from "@/projects/projects.module.scss";
-import star from "./starwars.module.scss";
+import styles from "@/scss/projects.module.scss";
+import star from "@scss/pages/starwars.module.scss";
 
 export default function StarWarsAPI() {
   return (
