+++ website/src/lib/scss/layout/root.scss
@@ -1,4 +1,4 @@
-@use "./globals.scss";
+@use "@/scss/globals.scss";
 
 * {
   box-sizing: border-box;
