+++ website/src/lib/scss/pages/home.module.scss
@@ -1,4 +1,4 @@
-@use "./globals.scss";
+@use "@/scss/globals.scss";
 
 $background-color_1: rgba(globals.$callout-rgb, 0.5);
 
@@ -84,31 +84,6 @@ $background-color_1: rgba(globals.$callout-rgb, 0.5);
     background-color: rgb(187, 180, 174);
     width: 100%;
     color: black;
-    .jobCard {
-      display: grid;
-      grid-template-areas:
-        "header header header"
-        "img desc desc"
-        "img desc desc";
-      padding: 1rem;
-      gap: 1rem;
-      width: 80%;
-      header {
-        grid-area: header;
-      }
-      .imgSection {
-        display: grid;
-        grid-template-columns: 1fr;
-        grid-area: img;
-        span {
-          max-width: 100%;
-          max-height: auto;
-        }
-      }
-      p {
-        grid-area: desc;
-      }
-    }
   }
 }
