+++ website/src/lib/scss/pages/starwars.module.scss
@@ -1,4 +1,4 @@
-@use "@/app/globals.scss";
+@use "@/scss/globals.scss";
 
 $starsURL: url("../../../../public/starry\ night\ -\ Transparent.png");
 $twinkleURL: url("../../../../public/twinkling.png");
