+++ website/src/lib/scss/projects.module.scss
@@ -1,4 +1,4 @@
-@use "@/app/globals.scss";
+@use "@/scss/globals.scss";
 
 .main {
   display: flex;
