+++ website/tsconfig.json
@@ -25,6 +25,7 @@
       "@/projects/*": ["src/app/projects/*"],
       "@/lib/*": ["src/lib/*"],
       "@/components/*": ["src/lib/components/*"],
+      "@/scss/*": ["src/lib/scss/*"],
       "@/data/*": ["src/lib/data/*"],
       "@/interface/*": ["src/lib/interface/*"],
       "@/util/*": ["src/lib/util/*"]
